+++ layabauth/_authentication.py
@@ -13,21 +13,19 @@ class User:
         self.name = oauth2helper.user_name(decoded_body)
 
 
-def authorizations(**scopes) -> dict:
+def authorizations(auth_url, **scopes) -> dict:
     """
     Return all security definitions.
-    Contains only one OAuth2 definition using Engie Azure authentication.
+    Contains only one OAuth2 implicit flow definition.
 
+    :param auth_url: Authorization URL.
     :param scopes: All scopes that should be available (scope_name = 'description as a string').
     """
-    engie_tenant_id = "24139d14-c62c-4c47-8bdd-ce71ea1d50cf"
-    nonce = scopes.pop("nonce", "7362CAEA-9CA5-4B43-9BA3-34D7C303EBA7")
-
     return {
         "oauth2": {
             "scopes": scopes,
             "flow": "implicit",
-            "authorizationUrl": f"https://login.microsoftonline.com/{engie_tenant_id}/oauth2/authorize?nonce={nonce}",
+            "authorizationUrl": auth_url,
             "type": "oauth2",
         }
     }